% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference.R
\name{load_reference}
\alias{load_reference}
\title{Load growth reference}
\usage{
load_reference(
  refcode = NULL,
  element = c("table", "index", "study", "all"),
  con = NULL
)
}
\arguments{
\item{refcode}{String, code of a reference. Only the first element is loaded.}

\item{element}{String, either \code{"table"}, \code{"index"}, \code{"study"} or
\code{"all"}.  The default is \code{"table"}.}

\item{con}{A database connection. The default \code{con = NULL} reads the
data from the internal data of the \code{jamesreferences} package.
Currently not implemented.}
}
\value{
The return value depends on the \code{element} parameter.\tabular{ll}{
   \code{element} \tab Return value \cr
   \code{"table"} \tab A \code{tibble} with reference values \cr
   \code{"index"} \tab Numeric vector with index values \cr
   \code{"study"} \tab Named character vector with study data \cr
   \code{"all"} \tab All stored information \cr
}
}
\description{
This function provides access to the built-in growth references.
}
\examples{
ref <- load_reference("gc_2019_dsc_male_")
}
