% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_refcode.R
\name{make_refcode}
\alias{make_refcode}
\title{Make reference code from components}
\usage{
make_refcode(name = "", year = "", yname = "", sex = "", sub = "")
}
\arguments{
\item{name}{Character vector identifying the reference name field}

\item{year}{Character vector identifying the reference year field}

\item{yname}{Character vector identifying the reference yname field}

\item{sex}{Character vector identifying the reference sex field}

\item{sub}{Character vector identifying the reference sub field}
}
\value{
A character vector with \code{n} elements, where \code{n} equals the longest
vector in the input arguments.
}
\description{
This function creates reference codes by pasting \code{name}, \code{year}, \code{yname},
\code{sex} and \code{sub} fields. The result is a string that identifies the
reference.
}
\note{
The function silently converts any zero-length arguments to length \code{1}.
The procedure aborts with an error if the length of any input is unequal to
either \code{1} or \code{n}.
}
\examples{
data <- data.frame(len = c(56, 42, 53),
                   age = c(0.1, 0.2, 0.15),
                   sex = c("male", "female", "female"),
                   ga = c(40, 27, 39))
make_refcode("nl", "1997", "hgt", data$sex, data$ga)
}
