% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_refcode_dts.R
\name{make_refcode_dts}
\alias{make_refcode_dts}
\title{Creates the refcode from dot arguments}
\usage{
make_refcode_dts(xname = "age", yname, x, sex = "", age = NULL, ga = NULL, ...)
}
\arguments{
\item{xname}{x}

\item{yname}{x}

\item{x}{x}

\item{sex}{x}

\item{age}{x}

\item{ga}{x}

\item{\dots}{not used}
}
\value{
Character vector with reference codes
}
\description{
Creates the refcode from dot arguments
}
