% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ref.R
\name{read_ref}
\alias{read_ref}
\title{Reads a REF formatted file into a tibble}
\usage{
read_ref(file)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \verb{http://},
\verb{https://}, \verb{ftp://}, or \verb{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. It must contain at
least one new line to be recognised as data (instead of a path) or be a
vector of greater than length 1.

Using a value of \code{\link[readr:clipboard]{clipboard()}} will read from the system clipboard.}
}
\value{
A tibble with an attribute called \code{"study"}.
}
\description{
Reads a file that is stored in REF format and returns a tibble. The function
validates the presence of required fields, and checks whether the variable
names in the data match the specified distribution.
}
\examples{
fn <- system.file("testdata/be_2009_bmi_female_.txt", package = "jamesreferences")
myref <- read_ref(fn)
}
\author{
Stef van Buuren 2021
}
