% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_refcodes.R
\name{set_refcodes}
\alias{set_refcodes}
\title{Calculates a name for a growth references for each measurement}
\usage{
set_refcodes(data)
}
\arguments{
\item{data}{Data frame in the long form, where each measurement has its
own row. Required variables are \code{xname}, \code{yname}, \code{x} and \code{sex}.
Optional variables are \code{age} and \code{ga}.}
}
\value{
A character vector of \code{nrow(data)} elements.
}
\description{
Determines the reference name for each measurement based on age, sex, outcome
and gestational age.
}
\details{
The present implementation uses three Dutch references: a term reference,
a pre-term reference for children born \code{ga <= 36} aged 0 - 4 years, and
a reference for the D-score. For head circumference, pre-term references up to
1.5 years are used, and term references thereafter. Together, this
specification defines the \strong{analysis metric} in JAMES.

The pre-term references govern gestational ages between 25 and 36 weeks.
The function scores pre-term born infants with gestational ages between
21 and 24 weeks relative to the 25-week references.

If the variable \code{ga} is not present in \code{data}, the function assumes
term births for all children and prints a message.

References for the Development Score (D-score) for pre-term were created
by shifting the term reference by 40 - ga weeks to the right, so the
references start at a later calender age. Because pre-terms are older
their Z-score initially score higher than those of terms. The effect
disappears after four months.

The \code{yname} field currently supports the following measurements:
\code{hgt} (length/height in cm), \code{wgt} (body weight in kg),
\code{hdc} (head circumference in cm), \code{bmi} (body mass index in kg/m**2),
\code{wfh} (weight for height in kg) and \code{dsc} (D-score).

The \code{xname} field currently supports the following measurements:
\code{age} (child age, decimal years) and \code{hgt} (height in cm, only for \code{wfh}).
}
\examples{
df <- data.frame(
  xname = c("age", "age", "age", "age", "hgt", "age"),
  yname = c("hgt", "hgt", "hdc", "hdc", "wfh", "dsc"),
  x = c(0.1, 0.1, 0.1, 1.6, 60, 1.0),
  sex = rep(c("male", "female"), 3),
  ga = c(39, 27, 27, 27, 39, 40))
refcodes <- set_refcodes(df)

# show the preterm 27 weeks hgt reference
centile::load_reference(refcodes[2], pkg = "jamesyzy")
}
\references{
\href{https://d-score.org/dbook1/sec-pops.html#conclusions}{D-score adjustment for pre-terms}
}
\seealso{
\code{\link[centile:make_refcode]{centile::make_refcode()}}, \code{\link[centile:load_reference]{centile::load_reference()}}
}
\author{
Stef van Buuren 2021
}
