% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform2y.R
\name{transform2y}
\alias{transform2y}
\title{Transforms Z-scores into measurements using Dutch references}
\usage{
transform2y(
  data,
  znames = c("hgt_z", "wgt_z", "hdc_z", "wfh_z", "bmi_z", "dsc_z"),
  pkg = "nlreferences",
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame with appropriate variables, at least
\code{sex} and \code{age}. The names of the measurements can
be one or more of: \code{hgt_z}, \code{wgt_z}, \code{hdc_z}, \code{wfh_z}, \code{bmi_z} and \code{dsc_z}.}

\item{znames}{Character vector containing the names of the Z-scores
to convert. Specify this to limit the number of conversions. If not specified, the
function calculates measurements for all Z-scores.}

\item{pkg}{Name of the package that stores the growth references. By
default, \code{transform2z} searches the \code{nlreferences} package.}

\item{verbose}{Set to \code{TRUE} to turn on warnings and messages, which is
useful for tracking down problem related to missing data or to the
availability of references.}
}
\value{
A data frame with either zero rows or the same number of rows
as \code{nrow(data)} with colums named \code{hgt}, \code{wgt}, and so on.
}
\description{
#' This function transforms Z-scores stored in the wide matrix
to measurements using the selector method implemented in
\code{\link[=set_refcodes]{set_refcodes()}}. It is the inverse of \code{\link[=transform2z]{transform2z()}}.
}
\details{
By default, the function scans for variables named
\code{hgt_z}, \code{wgt_z}, \code{hdc_z}, \code{wfh_z}, \code{bmi_z} and \code{dsc_z}, and returns
the measurement values for any variables it finds.

The conversion of \code{wfh_z} to \code{wfh} is special. As \code{wfh} is conditional
on \code{hgt}, the calculation requires either \code{hgt} or \code{hgt_z} to be present
in the data. Availability of \code{hgt} takes precedence over \code{hgt_z}. If
only \code{hgt_z} is known, then the function calculates \code{hgt} from \code{hgt_z}.
}
\examples{
df <- data.frame(hgt_z = 0, wgt_z = 1, hdc_z = -1, age = 0.3,
sex = "male", ga = c(20, 30, 40, 50))
transform2y(df, znames = c("hdc_z", "wfh_z"))
}
\author{
Stef van Buuren 2021
}
