% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform2z.R
\name{transform2z}
\alias{transform2z}
\title{Transforms measurements to Z-scores using Dutch references}
\usage{
transform2z(
  data,
  ynames = c("hgt", "wgt", "hdc", "wfh", "bmi", "dsc"),
  pkg = "nlreferences",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Data frame with appropriate variables, at least
\code{sex} and \code{age}. The names of the measurements can
be one or more of: \code{hgt}, \code{wgt}, \code{hdc}, \code{wfh}, \code{bmi} and \code{dsc}.}

\item{ynames}{Character vector containing the measurements to convert.
Specify this to limit the number of conversions. If not specified, the
function calculates Z-scores for all measurements.}

\item{pkg}{Name of the package that stores the growth references. By
default, \code{transform2z} searches the \code{nlreferences} package.}

\item{verbose}{Set to \code{TRUE} to turn on warnings and messages, which is
useful for tracking down problem related to missing data or to the
availability of references.}

\item{\dots}{Arguments passed down to \code{\link[centile:y2z]{centile::y2z()}}, e.g. \code{rule = c(1, 2)}
to calculate Z-scores beyond the maximum age in the reference table.}
}
\value{
A data frame with either zero rows or the same number of rows
as \code{nrow(data)} with colums named \code{hgt_z}, \code{wgt_z}, and so on.
}
\description{
This function transforms growth data in the wide matrix
to Z-scores using the selector method implemented in
\code{\link[=set_refcodes]{set_refcodes()}}. It is the inverse of \code{\link[=transform2y]{transform2y()}}.
}
\details{
By default, the function scans for variables named
\code{hgt}, \code{wgt}, \code{hdc}, \code{wfh}, \code{bmi} and \code{dsc}, and returns
Z-scores for any variables it finds.
}
\examples{
df <- data.frame(hgt = 60, wgt = 5, hdc = 40, age = 0.3,
sex = "male", ga = c(20, 30, 40, 50))
transform2z(df, ynames = c("hdc", "wfh"))
}
\seealso{
\code{\link[=set_refcodes]{set_refcodes()}}, \code{\link[centile:y2z]{centile::y2z()}}
}
\author{
Stef van Buuren 2021
}
