% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_z.R
\name{transform_z}
\alias{transform_z}
\title{Transforms measurement to Z-scores using Dutch references}
\usage{
transform_z(
  data,
  ynames = c("hgt", "wgt", "hdc", "wfh", "bmi", "dsc"),
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame with appropriate variables, at least
\code{sex} and \code{age}. The names of the measurements can
be one or more of: \code{hgt}, \code{wgt}, \code{hdc}, \code{wfh}, \code{bmi} and \code{dsc}.
If the variable \code{ga} is entirely missing, the function assumes term
births and prints a message.}

\item{ynames}{Character vector containing the measurements to convert.
Specify this to limit the number of conversions. If not specified, the
functions tries all transformations}

\item{verbose}{Set to \code{TRUE} to turn on warnings and messages}
}
\value{
A data frame with either zero rows or the same number of rows
as \code{nrow(data)} with colums named \code{hgt_z}, \code{wgt_z}, and so on.
}
\description{
This function transforms growth data to Z-scores using two
Dutch references: a term reference and a preterm-reference for
children born \code{ga <= 36} aged 0 - 4 years.
For head circumference, preterm references up to
1.5 years are used, and term references thereafter.
}
\details{
By default, the functions scans for growth variables, and converts
any variables it finds.
}
\examples{
df <- data.frame(hgt = 60, wgt = 5, hdc = 40, age = 0.3,
sex = "male", ga = c(20, 30, 40, 50))
transform_z(df, ynames = c("hdc", "wfh"))
}
