% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\name{z}
\alias{z}
\title{Convert measurements into Z-scores}
\usage{
z(y, x, refcode, dec = 3L, tail_adjust = FALSE, ...)
}
\arguments{
\item{y}{A numerical vector with measurements`.}

\item{x}{A vector containing \code{length(y)} values of the numerical covariate (typically
decimal age or height) at which conversion is desired.}

\item{refcode}{A character vector with \code{length(y)} elements, each of which
points to a reference.}

\item{dec}{A scalar value indicating the number of decimals used to round the
value. The default is 3 decimals.}

\item{tail_adjust}{Logical. If \code{TRUE} the procedure applies the WHO
method for tail adjustment under the LMS distribution.
The default is \code{FALSE}.}

\item{\dots}{Not used.}
}
\value{
A vector with \code{length(y)} elements containing the Z-scores.
}
\description{
Converts measurements into age- or hgt-conditional standard deviation scores
(SDS, Z-score) using one or more external reference distributions.
}
\note{
The type of reference distribution is defined by the \code{distribution} field
of the \code{study} attribute in the reference. The function executes
transformations specified by the \code{tx} and \code{ty} fields before calculating the
Z-score.

Functions \code{z()} and \code{y()} functions are the inverse of each other.
}
\examples{
y <- c(rep(c(40, 5), 4))
x <- rep(c(0.3, 60), 4)
refcode <- c(
  "nl_2012_hdc_male_20", "nl_2012_wfh_male_",
  "nl_2012_hdc_male_30", "nl_2012_wfh_male_",
  "nl_1997_hdc_male_nl", "nl_1997_wfh_male_nla",
  "nl_1997_hdc_male_nl", "nl_1997_wfh_male_nla"
)
z <- z(y, x, refcode)
}
\author{
Stef van Buuren, 2021
}
